[Refactoring] Split ConvertToComputedProperty to its own file
